优化修改指令，更新README.md

+++ src/main/java/world/thek/controller/OrderController.java
@@ -9,9 +9,10 @@ import world.thek.util.FileUtil;
 
 import java.io.IOException;
 import java.util.Map;
+import java.util.Objects;
 
 /**
- * @author intasect
+ * @author thek
  * @date:  2022/9/21 上午11:05
  */
 public class OrderController extends SimpleListenerHost {
@@ -21,10 +22,10 @@ public class OrderController extends SimpleListenerHost {
         String[] split = code.split("\\s+");
         String order = split[0];
 
-        Map<String,String> studyMap = FileUtil.read(ConstantUtil.ORDER_FILENAME);
+        Map<String,String> orderMap = FileUtil.read(ConstantUtil.ORDER_FILENAME);
 
-        if (studyMap.get(order) != null) {
-            String value = studyMap.get(order);
+        if (orderMap.get(order) != null) {
+            String value = orderMap.get(order);
             event.getSubject().sendMessage(value);
         }
 
@@ -32,7 +33,7 @@ public class OrderController extends SimpleListenerHost {
         if (Order.ORDER_ADD.equals(order)) {
             String key = split[1];
             String value = split[2];
-            if (studyMap.get(key) != null) {
+            if (orderMap.get(key) != null) {
                 event.getSubject().sendMessage("指令已存在！");
             } else {
                 FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
@@ -43,8 +44,8 @@ public class OrderController extends SimpleListenerHost {
         //查询指令
         if (Order.ORDER_FIND.equals(order)) {
             String key = split[1];
-            if (studyMap.get(key) != null) {
-                String value = studyMap.get(key);
+            if (orderMap.get(key) != null) {
+                String value = orderMap.get(key);
                 event.getSubject().sendMessage(key+" "+value);
             } else {
                 event.getSubject().sendMessage("不存在指令："+key);
@@ -56,15 +57,19 @@ public class OrderController extends SimpleListenerHost {
         if (Order.ORDER_UPDATE.equals(order)) {
             String key = split[1];
             String value = split[2];
-            FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
-            FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
-            event.getSubject().sendMessage("修改成功："+key);
+            if (orderMap.get(key) != null) {
+                FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
+                FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
+                event.getSubject().sendMessage("修改成功："+key);
+            } else {
+                event.getSubject().sendMessage("不存在指令："+key);
+            }
         }
 
         //所有指令
         if(Order.ORDER_ALL.equals(order)) {
             String all = FileUtil.findALL(ConstantUtil.ORDER_FILENAME);
-            if (all != "") {
+            if (!Objects.equals(all, "")) {
                 event.getSubject().sendMessage(all);
             } else {
                 event.getSubject().sendMessage("暂无指令");
@@ -74,8 +79,7 @@ public class OrderController extends SimpleListenerHost {
         //删除指令
         if (Order.ORDER_REMOVE.equals(order)) {
             String key = split[1];
-            if (studyMap.get(key) != null) {
-                String value = studyMap.get(key);
+            if (orderMap.get(key) != null) {
                 FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
                 event.getSubject().sendMessage("已删除："+key);
             } else {
